# configuration-server-api/src/main/java/es/gob/radarcovid/configuration/api/TextCustomMap.java
/*
 * Copyright (c) 2020 Gobierno de España
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 *
 * SPDX-License-Identifier: MPL-2.0
 */
package es.gob.radarcovid.configuration.api;

import java.io.Serializable;
import java.util.TreeMap;

public class TextCustomMap extends TreeMap<String, String> implements Serializable {

}